requirements.txt
beautifulsoup4==4.9.1
lxml==4.5.2
selenium==3.141.0
soupsieve==2.0.1
urllib3==1.25.10
beautifulsoup4==4.9.1
lxml==4.5.2
python-dotenv==0.14.0
selenium==3.141.0
soupsieve==2.0.1
urllib3==1.25.10
beautifulsoup4==4.9.1
flake8==3.8.3
lxml==4.5.2
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.14.0
selenium==3.141.0
soupsieve==2.0.1
urllib3==1.25.10
